Add support for dynamic BAT modification, part 1.
This is the easy part: making things work in interpreter. Also gets rid of
the BAT option; real BAT lookups into the precomputed table should be cheap
enough that it isn't worth keeping an option around. (If the impact is
actually measurable, we could do something fancier like a table of
function pointers.)